Update parameters and remove unused code in partition_iterator_state_based.py
The commit updates the descriptions and datatypes for the param variables 'alias_path_data' and 'alias_path_outputs' in PartitionIterator initialization. Additionally, a line of code related to 'delete_feature_class' property was removed due to redundancy. Also, the 'output' key in the 'outputs' dictionary was replaced with 'input', and 'building_polygons' key's value was commented out.