Refactor partition_io_decorator location and add `define_io_params` method
Moved the `partition_io_decorator` from the class level to the method level in the `PolygonProcessor` class. Also, in the `partition_iterator.py` file, a new method `define_io_params` was added to the `run` method. These modifications bring increased clarity and structure to the code by specifying the scope of the decorator and adding auxiliary functionality to define I/O parameters.