fixed bug in begrensningskurve_land_waterbodies.py where polygon features was present on land in final output
fixed bug in begrensningskurve_land_waterbodies.py where polygon features was present on land in final output. This commit also refines the buffer calculation for the water boundary by including the water barrier buffer width in it. A method 'delete_feature_class' is also updated to be a static method.