Removed point to polygon squares logic from resolve_building_conflicts.py to outside the logic.
While hunting the RBC bug we have moved the logic creating the building point square creation to outside the RBC logic prior on the global data in the point_resolve_building_conflicts.py. There are some syntax testing changes to RBC which will probably not remain as we are testing for the source of the bug inside Esri's tool.