Add initial key-value pair when creating a new alias
Previously, newly created aliases were initialized with empty dictionaries, which resulted in problems when appending new alais to final outputs for partition_iterator.py This fixes this issue. However, this solution assume that as long as one input feature is present for the partition that it is ok to greenlight new alias outputs for appending to final outputs. Edge cases need to be handled.