Thin roads cycle update (#306)
* WIP on iterating thin_road_network

Started working on logic to make thin roads iterate untill no more roads
are removed.

* Fixed dissolving fields inside partition_iterator.

When using thin_road_network inside partition_iterator it called
DissolveWithIntersections which takes fields to keep as a param. Needed
to make it possible to inject the generated partition_field_name from
partition_iterator as a param for classes and functions.

* Moved count_objects function to file_utilities

Instead of having count_objects function as local functions in classes
created a function in file_utilities instead.

* Added working partition size logic for partition_iterator

Now it is a working logic to determine the partition size used based on
feature_count and search_distance. Now the class iterates down until the
used feature_count is not exceeding the limit.

* Small partition_iterator fix and road run optimization.

Fixed partition_iterator to default to input feature_count if
run_partition_optimization is set to false. Also removed redundant code
in data_preparation_2, and other optimization.

* Added logic to partition_iterator to skip testing identical partitions

* Added Integrate to data_preparation_2

* Removed integrate and remove small lines

* Changing resolve_road_conflicts partition partition size

* removed obsolete resolve_road_conflict